Jochris

+++ Assets/EnglishBattleRoyale/Scripts/Interfaces/IQuestionTimeObserver.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections;
 public interface IQuestionTimeObserver {
-	void OnStartQuestionTimer (Action<int> action, int timer);
+	void OnStartQuestionTimer (Action<float> action, float timer);
 	void OnStopQuestionTimer ();
-
 }
